[Feat] #272 - 메뉴 추가와 수정 ViewModel 분리
