+++ Hankkijogbo/Hankkijogbo/Present/HankkiMenu/ViewModel/AddMenuViewModel.swift
@@ -48,22 +48,4 @@ extension AddMenuViewModel {
             }
         }
     }
-    
-    /// 메뉴 수정
-    func modifyMenuAPI(storeId: Int, id: Int, requestBody: [MenuData], completion: @escaping () -> Void) {
-        NetworkService.shared.menuService.patchMenu(storeId: storeId, id: id, requestBody: requestBody) { result in
-            result.handleNetworkResult { _ in
-                completion()
-            }
-        }
-    }
-    
-    /// 메뉴 삭제
-    func deleteMenuAPI(storeId: Int, id: Int, completion: @escaping () -> Void) {
-        NetworkService.shared.menuService.deleteMenu(storeId: storeId, id: id) { result in
-            result.handleNetworkResult { _ in
-                completion()
-            }
-        }
-    }
 }
